refactor: nest VariantData inside MergedBpmnModel and inline variants mapping
- VariantData is now MergedBpmnModel.VariantData (nested class)
- Inlined variants val directly into MergedBpmnModel constructor
